Fix the derivative of semi-constants
During initialisation, a constant, say y, might appear
differentiated in an equation's iteration variables.
Therefore, we need to set the partial derivatives correctly.

We do so, by setting relativeIndex to -n where n is the lowest
derivative of the corresponding variable in the iteration tuple.
In case of a real constant use a large enough n.

Signed-off-by: Christoph Höger <christoph.hoeger@tu-berlin.de>