Add inline integration
Since a.c.m integrators seem to produce quite some overhead
(and we're going to implement event handlers on our own
anyway), it seems to be a good idea to implement at least the
simple integrators directly. In some cases this bumps performance
by a factor of two (while in other cases it actually decreases
performance).

Signed-off-by: Christoph Höger <christoph.hoeger@tu-berlin.de>